requirements.txt
flask
gunicorn
scikit-learn==1.4.2
numpy==1.26.4
joblib
google-cloud-storage